WebUITestFrame/requirements.txt
allure_python_commons==2.13.2
imageio==2.31.2
numpy==1.21.6
opencv_python==4.9.0.80
Pillow==9.5.0
PyAutoGUI==0.9.54
pyperclip==1.8.2
pyscreenshot==3.1
pytesseract==0.3.10
pytest==7.4.0
PyYAML==6.0
Requests==2.31.0
selenium==4.17.2
